pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skill-scanner"
version = "0.1.0"
description = "Detect security threats in agent skill files (OpenClaw, MCP)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
skill-scanner = "skill_scanner.cli:main"

[tool.setuptools.packages.find]
include = ["skill_scanner*"]
